+++ src/game_object/scenic.py
